#5210: Get addcdiv stuck tests back to non working folder
