+++ tests/tt_eager/python_api_testing/non_working_unit_tests/grayskull/test_eltwise_addcdiv_stuck_1.py
@@ -4,6 +4,7 @@
 
 import random
 from loguru import logger
+import pytest
 import torch
 import tt_lib as ttl
 
@@ -19,10 +20,8 @@ def run_addcdiv(input_shape, dtype, dlayout, buffer_type, output_mem_config, dat
     x = torch.Tensor(size=input_shape).uniform_(-100, 100)
     y = torch.Tensor(size=input_shape).uniform_(-100, 100)
     z = torch.Tensor(size=input_shape).uniform_(-100, 100)
-    x = torch.where(y.abs() > 1e-3, x, 1e-3)
-    y = torch.where(y.abs() > 1e-3, y, 1e-3)
-    z = torch.where(y.abs() > 1e-3, z, 1e-3)
 
+    # ref_value = torch.addcdiv(x, y, z, value=scalar)
     logger.info(
         f"Running addcdiv with input_shape {input_shape} dtype {dtype} dlayout {dlayout} buffer_type {buffer_type} output_mem_config {output_mem_config} scalar {scalar} data_seed {data_seed}"
     )
@@ -54,6 +53,7 @@ def run_addcdiv(input_shape, dtype, dlayout, buffer_type, output_mem_config, dat
         t3 = ttl.tensor.addcdiv(t0, t1, t2, scalar, output_mem_config)
 
         y = tt2torch_tensor(t3)
+
     except Exception as exc:
         logger.warning(f"run_addcdiv RuntimeError occured {exc}")
 
@@ -63,43 +63,40 @@ def run_addcdiv(input_shape, dtype, dlayout, buffer_type, output_mem_config, dat
     logger.info(f"Finished running addcdiv")
 
 
-def test_addcdiv_test():
-    run_addcdiv(
-        (3, 10, 73, 388),
-        dtype=[ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT16],
-        dlayout=[ttl.tensor.Layout.ROW_MAJOR, ttl.tensor.Layout.ROW_MAJOR, ttl.tensor.Layout.ROW_MAJOR],
-        buffer_type=[ttl.tensor.BufferType.DRAM, ttl.tensor.BufferType.DRAM, ttl.tensor.BufferType.DRAM],
-        output_mem_config=ttl.tensor.MemoryConfig(
-            ttl.tensor.TensorMemoryLayout.INTERLEAVED, ttl.tensor.BufferType.DRAM
-        ),
-        data_seed=8405597,
-        scalar=-61.75,
-    )
-
-    run_addcdiv(
+test_sweep_args = [
+    (
         (9, 23, 416, 310),
-        dtype=[ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT16],
-        dlayout=[ttl.tensor.Layout.ROW_MAJOR, ttl.tensor.Layout.ROW_MAJOR, ttl.tensor.Layout.ROW_MAJOR],
-        buffer_type=[ttl.tensor.BufferType.DRAM, ttl.tensor.BufferType.DRAM, ttl.tensor.BufferType.DRAM],
-        output_mem_config=ttl.tensor.MemoryConfig(
-            ttl.tensor.TensorMemoryLayout.INTERLEAVED, ttl.tensor.BufferType.DRAM
-        ),
-        data_seed=10406825,
-        scalar=-42.25,
-    )
-
-    run_addcdiv(
+        [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT16],
+        [ttl.tensor.Layout.ROW_MAJOR, ttl.tensor.Layout.ROW_MAJOR, ttl.tensor.Layout.ROW_MAJOR],
+        [ttl.tensor.BufferType.DRAM, ttl.tensor.BufferType.L1, ttl.tensor.BufferType.DRAM],
+        ttl.tensor.MemoryConfig(ttl.tensor.TensorMemoryLayout.INTERLEAVED, ttl.tensor.BufferType.L1),
+        10406825,
+        -42.25,
+    ),
+    (
+        (3, 10, 73, 388),
+        [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT16],
+        [ttl.tensor.Layout.ROW_MAJOR, ttl.tensor.Layout.ROW_MAJOR, ttl.tensor.Layout.ROW_MAJOR],
+        [ttl.tensor.BufferType.DRAM, ttl.tensor.BufferType.DRAM, ttl.tensor.BufferType.L1],
+        ttl.tensor.MemoryConfig(ttl.tensor.TensorMemoryLayout.INTERLEAVED, ttl.tensor.BufferType.L1),
+        8405597,
+        -61.75,
+    ),
+    (
         (2, 24, 39, 462),
-        dtype=[ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT16],
-        dlayout=[ttl.tensor.Layout.ROW_MAJOR, ttl.tensor.Layout.ROW_MAJOR, ttl.tensor.Layout.ROW_MAJOR],
-        buffer_type=[
-            ttl.tensor.BufferType.DRAM,
-            ttl.tensor.BufferType.DRAM,
-            ttl.tensor.BufferType.DRAM,
-        ],  # too large for L1
-        output_mem_config=ttl.tensor.MemoryConfig(
-            ttl.tensor.TensorMemoryLayout.INTERLEAVED, ttl.tensor.BufferType.DRAM
-        ),  # too large for L1
-        data_seed=10406825,
-        scalar=-42.25,
-    )
+        [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT16],
+        [ttl.tensor.Layout.ROW_MAJOR, ttl.tensor.Layout.ROW_MAJOR, ttl.tensor.Layout.ROW_MAJOR],
+        [ttl.tensor.BufferType.L1, ttl.tensor.BufferType.DRAM, ttl.tensor.BufferType.L1],
+        ttl.tensor.MemoryConfig(ttl.tensor.TensorMemoryLayout.INTERLEAVED, ttl.tensor.BufferType.L1),
+        10406825,
+        -42.25,
+    ),
+]
+
+
+@pytest.mark.parametrize(
+    "input_shape, dtype, dlayout, buffer_type, output_mem_config, data_seed, scalar",
+    (test_sweep_args),
+)
+def test_addcdiv_test(input_shape, dtype, dlayout, buffer_type, output_mem_config, data_seed, scalar):
+    run_addcdiv(input_shape, dtype, dlayout, buffer_type, output_mem_config, data_seed, scalar)
